Move shared modules to designated folder (#971)
