+++ lib/dialogs/index.js
@@ -1,5 +1,3 @@
 export { default as About } from './about';
 export { default as Settings } from './settings';
 export { default as Share } from './share';
-
-export { default as Types } from './types';
