+++ lib/note-toolbar-container.js
@@ -6,7 +6,7 @@ import { get, includes } from 'lodash';
 import analytics from './analytics';
 import appState from './flux/app-state';
 import { toggleFocusMode } from './state/settings/actions';
-import DialogTypes from './dialogs/types';
+import DialogTypes from '../shared/dialog-types';
 import filterNotes from './utils/filter-notes';
 
 export class NoteToolbarContainer extends Component {
