+++ shared/dialog-types.js
